utilize ts much more in the redux setup
