+++ package.json
@@ -14,12 +14,14 @@
     "react-router-dom": "^4.2.2",
     "react-router-redux": "^4.0.8",
     "react-scripts-ts": "2.14.0",
+    "redux": "3.7.2",
     "redux-devtools-extension": "^2.13.2",
     "redux-thunk": "^2.2.0",
     "save": "^2.3.2",
     "semantic-ui-css": "^2.3.1",
     "semantic-ui-react": "^0.79.1",
-    "typesafe-actions": "^1.1.2"
+    "typesafe-actions": "^1.1.2",
+    "utility-types": "1.1.0"
   },
   "scripts": {
     "predeploy": "npm run build",
@@ -34,7 +36,8 @@
     "@types/node": "^9.4.7",
     "@types/react": "^16.0.40",
     "@types/react-dom": "^16.0.4",
-    "typescript": "^2.7.2"
+    "typescript": "^2.7.2",
+    "tslint-react": "3.3.3"
   },
   "homepage": "https://esow.github.io/buddy_v2"
 }
