+++ src/store/model/auth.ts
@@ -1,0 +1,4 @@
+export interface AuthSessionDTO {
+	session_token: String;
+	session_id: String;
+}
