+++ tsconfig.json
@@ -1,11 +1,12 @@
 {
   "compilerOptions": {
     "outDir": "build/dist",
-    "module": "esnext",
-    "target": "es5",
+    "module": "es2015",
+    "target": "es2015",
     "lib": [
-      "es6",
-      "dom"
+      "dom",
+      "es2016",
+      "es2017.object"
     ],
     "sourceMap": true,
     "allowJs": true,
